+++ o3d3xx-o3x1xx/multi_camera_grabber/ex-multi_camera_grabber.cpp
@@ -1,17 +1,6 @@
 /*
- * Copyright (C) 2018 ifm electronics, gmbh
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distribted on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright 2022-present ifm electronic, gmbh
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 //
@@ -19,12 +8,12 @@
 //
 // Capture frames from multiple ifm 3D cameras which are configured to be triggered through software,
 // and display the time stamp of the frame received.
-// also measues the time taken to receive the set of frames.
+// also measure the time taken to receive the set of frames.
 //
 // Prerequisites:
-// *) Each ifm 3D camera should be configured to use "Process Interface" for trigger.
-// *) You Should be able to ping each of the 3D camera from the PC on which this code executes.
-// *) Incase your network uses a proxy, you will need to configure your system to bypass the proxy for the used IP's.
+// 1) Each ifm 3D camera should be configured to use "Process Interface" for trigger.
+// 2) You Should be able to ping each of the 3D camera from the PC on which this code executes.
+// 3) Incase your network uses a proxy, you will need to configure your system to bypass the proxy for the used IP's.
 //
 
 #include <iostream>
@@ -33,19 +22,20 @@
 #include <vector>
 #include <ctime>
 #include <iomanip>
-#include <ifm3d/camera.h>
+#include <sstream>
+
+
+#include <ifm3d/device/device.h>
+#include <ifm3d/device/legacy_device.h>
 #include <ifm3d/fg.h>
-#include <ifm3d/image.h>
 
 namespace
 {
-    //CHANGE IP addreses to those of camera's avaialible!!!!
-    const auto CAMERA0 = "192.168.0.70";
-    const auto CAMERA1 = "192.168.0.71";
-    const auto CAMERA2 = "192.168.0.72";
-
-    //Add the IP of cameras to be used.
-    const std::array<const std::string,3> camera_ips = {CAMERA0, CAMERA1, CAMERA2};
+    // List of camera devices to be used for grabbing.
+    std::vector<std::string> camera_ip_list = {
+        "192.168.0.68",
+        "192.168.0.80"
+    };
 
     //Utility function to format the timestamp
     std::string formatTimestamp(ifm3d::TimePointT timestamp)
@@ -66,55 +56,63 @@ namespace
     }
 }
 
-
 int main(int argc, const char **argv)
 {
     std::chrono::system_clock::time_point start, end;
 
-    //vectors for the objects to be used.
-    std::vector<ifm3d::Camera::Ptr> cameras;
-    std::vector<ifm3d::FrameGrabber::Ptr> frame_grabbers;
-    std::vector<ifm3d::ImageBuffer::Ptr> image_buffers;
+    //vector of objects
+    std::vector<ifm3d::Device::Ptr> devices;
+    std::vector<ifm3d::FrameGrabber::Ptr> grabbers;
+    ifm3d::json config ;
 
-    // Create ifm3d objects of Camera, ImageBuffer and FrameGrabber for each of the camera devices.
-    for(auto camera_ip:camera_ips)
+    for (const auto& camera_ip : camera_ip_list)
     {
-        auto cam = ifm3d::Camera::MakeShared(camera_ip);
-        cameras.push_back(cam);
-        image_buffers.push_back(std::make_shared<ifm3d::ImageBuffer>());
-        frame_grabbers.push_back(std::make_shared<ifm3d::FrameGrabber>(cam));
+        auto cam = ifm3d::Device::MakeShared(camera_ip);
+        devices.push_back(cam);
 
+        // Create a frame grabber object
+        grabbers.push_back(std::make_shared<ifm3d::FrameGrabber>(cam));
+        auto legacy_device = std::dynamic_pointer_cast<ifm3d::LegacyDevice>(cam);
+       
+        //mark the current active application as sw triggered
+        int idx = legacy_device->ActiveApplication();
+        config = legacy_device->ToJSON();
+        config["ifm3d"]["Apps"][idx - 1]["TriggerMode"] =
+        std::to_string(static_cast<int>(ifm3d::Device::trigger_mode::SW));
+        legacy_device->FromJSON(config);
     }
 
-    int count = 0;
-    while (count++ < 10)
+    // grab the frames from the multiple cameras
+    int frame_count = 10;
+    for(int i=0;i<frame_count;i++)
     {
-        //for each of the camera device, software trigger is sent and wait for frame is done sequentially.
         start = std::chrono::system_clock::now();
-        for(int index = 0; index<camera_ips.size();index++ )
+        uint8_t id = 0;
+        for(const auto& camera_ip : camera_ip_list)
         {
-            auto frame_grabber = frame_grabbers.at(index);
-            auto image_buffer = image_buffers.at(index);
-
-            frame_grabber->SWTrigger();
-            if (!frame_grabber->WaitForFrame(image_buffer.get(), 1000))
+            auto fg = grabbers.at(id);
+            // use SWtrigger
+            std::cout << "SW trigger for camera(" << camera_ip << ")" << std::endl;
+            fg->Start({ifm3d::buffer_id::AMPLITUDE_IMAGE});
+            fg->SWTrigger();
+            auto frame = fg->WaitForFrame();
+            if(frame.wait_for(std::chrono::milliseconds(1000)) != std::future_status::ready)
             {
-                std::cerr << "Timeout waiting for camera("<<camera_ips[index]<<") !" << std::endl;
+                std::cerr << "timeout waiting for camera(" << camera_ip << ") frame" << std::endl;
                 return -1;
             }
-            else {
-                ifm3d::TimePointT timestamp = image_buffer->TimeStamp();
-                std::cout << "got camera("<<camera_ips[index]<<") frame, timestamp"
-                          << std::setw(2) << std::setfill('0')
-                          << ": " << formatTimestamp(timestamp)
-                          << std::endl;
+            else
+            {
+                ifm3d::TimePointT timestamp = frame.get()->TimeStamps().front();
+                std::cout << "got camera(" << camera_ip << ") frame timestamp "
+                << std::setw(2) << std::setfill('0')
+                << ":" << formatTimestamp(timestamp) << std::endl;
             }
-
+            id++;
         }
         end = std::chrono::system_clock::now();
-        std::chrono::duration<double, std::milli> duration_ms = (end - start);
-        std::cout << "total get time in ms: " << duration_ms.count()<< std::endl;
+        std::chrono::duration<double,std::milli> duration_ms = (end -start);
+        std::cout << "total time taken to receive in ms " << duration_ms.count() << " ms" << std::endl;
     }
-
-    return 0;
+return 0;
 }